+++ CHANGELOG.md
@@ -620,3 +620,5 @@
 * Define `Same` type alias
 * Define `Same` pattern synonym
 * Define `State` pattern synonym
+* Define `o_yoklKL` operator
+* Define `oo_yoklKL` operator